requirements.txt
Flask>=2.0
Flask-Mail>=0.9.1
python-dotenv>=1.0
gunicorn>=20.1.0
